requirements.txt
audiomentations>=0.30.0
Cython>=0.29.33
fugashi>=1.2.1
huggingface-hub>=0.12.0
hydra-core>=1.3.1
jaconv>=0.3.3
japanize-matplotlib>=1.1.3
joblib>=1.2.0
librosa>=0.9.2
matplotlib>=3.6.3
pandas>=1.5.3
pydantic>=1.10.7
pyopenjtalk>=0.3.0
python-dotenv>=0.21.1
pytest>=7.3.1
pytorch-lightning>=2.0.2
scikit-learn>=1.2.1
scipy>=1.10.0
soundfile>=0.11.0
sox>=1.4.1
tensorboard>=2.11.2
timm>=0.6.12
torch @ https://download.pytorch.org/whl/cu118/torch-2.0.0%2Bcu118-cp310-cp310-linux_x86_64.whl
torchaudio @ https://download.pytorch.org/whl/cu118/torchaudio-2.0.1%2Bcu118-cp310-cp310-linux_x86_64.whl
torchmetrics>=0.11.1
torchvision @ https://download.pytorch.org/whl/cu118/torchvision-0.15.1%2Bcu118-cp310-cp310-linux_x86_64.whl
tqdm>=4.64.1
transformers>=4.25.1
wasabi>=0.10.1

adan @ git+https://github.com/sail-sg/Adan.git@8362d90a8f8f9c9315177d193c24094b1a610308